Updated Docker scripts, fixed master name detection, buildlocal, Jenkinsfile
Added new features to Docker distribution. "FailIfPull" marks the worker as
failed if the image needs to be pulled, and pulls it in the background
meanwhile. The distribution tool now looks up if the image exists in the
DockerHub or private registry if the "--pull" flag is set (although it adds a
little overhead for the checking). Modes have been added to the distribution
tools, used to express the preference: Fast execution but no guarantee that it
will work at all, or slower execution, making all the needed checkings, that
will most certainly work.

buildlocal script has been modified to fix the LXC base image creation.

Master name detection has been moved to a different part.

Jenkinsfile has been fixed. Hopefully it will work now.