Fix header_setup/replace_header.sh
replace_header.sh was removing any applied licence to add BSC one to the
source files. This is wrong and should not be done as it would strip
intellectual property of any external contribution.

The template is now broken in three parts. Any already applied copyright
is kept. Any file without a copyright are considered to be under the BSC
licencing to keep the previous behaviour.

The process of looking for extra copyrights is only applied to the first
header encountered (that is the job of the `extract_header_$lang.awk`
scripts) if any. The first block has to be a Apache Licence, Version 2.0
or there should not be any copyright block.

WARNING: Any file that are using another licencing or another
presentation will be stripped of it to have the template applied. Hence,
there is a risk of losing the licence if a different than Apache Licence
Version 2.0 is used.